refactor(app): move subtle files in root src folder inside to their respective folders
