+++ packages/app/src/App.tsx
@@ -1,13 +1,13 @@
 import { QueryClientProvider } from '@tanstack/react-query'
 import { type FC, useEffect } from 'react'
 
-import FileDropzone from './FileDropzone.tsx'
-import Scene from './Scene'
-import Sidebar from './Sidebar'
-import ToastContainer from './ToastContainer.tsx'
+import FileDropzone from './components/FileDropzone'
 import LeftButtonPanel from './components/LeftButtonPanel'
 import MobileBottomButtonPanel from './components/MobileBottomButtonPanel'
 import QuickActionPanel from './components/QuickActionPanel.tsx'
+import Scene from './components/Scene'
+import Sidebar from './components/Sidebar'
+import ToastContainer from './components/ToastContainer'
 import BlockDisplaySelectDialog from './components/dialog/BlockDisplaySelectDialog'
 import ExportToMinecraftDialog from './components/dialog/ExportToMinecraftDialog'
 import ItemDisplaySelectDialog from './components/dialog/ItemDisplaySelectDialog'
