+++ packages/app/src/components/LeftButtonPanel.tsx
@@ -14,10 +14,10 @@ import { useShallow } from 'zustand/shallow'
 import { newProject } from '@/lib/actions'
 import { openFileFromUserSelect, saveToFile } from '@/lib/file-handler'
 import { getLogger } from '@/lib/logger'
+import { getFormattedShortcutKeyString } from '@/lib/utils'
 import { useDialogStore } from '@/stores/dialogStore'
 import { useEditorStore } from '@/stores/editorStore'
 import { useHistoryStore } from '@/stores/historyStore'
-import { getFormattedShortcutKeyString } from '@/utils'
 
 import FloatingButton from './FloatingButton'
 import MobileDragHoldButton from './MobileDragHoldButton'
