+++ packages/app/src/components/SidePanel.tsx
@@ -1,6 +1,6 @@
 import { type JSX, forwardRef } from 'react'
 
-import { cn } from '@/utils'
+import { cn } from '@/lib/utils'
 
 const SidePanel = forwardRef<HTMLDivElement, JSX.IntrinsicElements['div']>(
   ({ className, ...props }, ref) => {
@@ -8,7 +8,7 @@ const SidePanel = forwardRef<HTMLDivElement, JSX.IntrinsicElements['div']>(
       <div
         {...props}
         className={cn(
-          'flex select-none flex-col gap-1 rounded-lg bg-neutral-900 p-2 text-sm',
+          'flex flex-col gap-1 rounded-lg bg-neutral-900 p-2 text-sm select-none',
           className,
         )}
         ref={ref}
