+++ packages/app/src/components/brandings.tsx
@@ -1,7 +1,7 @@
 import type { FC } from 'react'
 import { Trans, useTranslation } from 'react-i18next'
 
-import { cn } from '@/utils'
+import { cn } from '@/lib/utils'
 
 interface TitleProps {
   className?: string
