+++ packages/app/src/components/dialog/settings/ShortcutsPage.tsx
@@ -12,8 +12,8 @@ import { useTranslation } from 'react-i18next'
 import { LuCheck, LuEraser, LuUndo } from 'react-icons/lu'
 
 import type { Settings } from '@/lib/settings'
+import { cn, getFormattedShortcutKeyString } from '@/lib/utils'
 import { useEditorStore } from '@/stores/editorStore'
-import { cn, getFormattedShortcutKeyString } from '@/utils'
 
 const SPECIAL_KEYS = ['Control', 'Alt', 'Shift']
 const BLOCKED_KEYS = ['Unidentified']
