+++ packages/app/src/components/sidebar/ObjectsPanel.tsx
@@ -5,9 +5,9 @@ import { LuChevronRight, LuSmile, LuType } from 'react-icons/lu'
 import { TbDiamondFilled } from 'react-icons/tb'
 import { useShallow } from 'zustand/shallow'
 
+import { cn } from '@/lib/utils'
 import { useDisplayEntityStore } from '@/stores/displayEntityStore'
 import { isItemDisplayPlayerHead } from '@/types'
-import { cn } from '@/utils'
 
 import { SidePanel, SidePanelContent, SidePanelTitle } from '../SidePanel'
