+++ packages/app/src/components/ui/DropdownMenu.tsx
@@ -2,7 +2,7 @@ import * as DropdownMenuPrimitive from '@radix-ui/react-dropdown-menu'
 import * as React from 'react'
 import { LuCheck, LuChevronRight, LuCircle } from 'react-icons/lu'
 
-import { cn } from '@/utils'
+import { cn } from '@/lib/utils'
 
 const DropdownMenu = DropdownMenuPrimitive.Root
