+++ packages/app/src/components/ui/MultiSegmentProgress.tsx
@@ -1,7 +1,7 @@
 import * as ProgressPrimitive from '@radix-ui/react-progress'
 import * as React from 'react'
 
-import { cn } from '@/utils'
+import { cn } from '@/lib/utils'
 
 const REMAINING_AREA_NAME = '__remaining_area__'
 
@@ -46,7 +46,7 @@ const MultiSegmentProgress = React.forwardRef<
     <ProgressPrimitive.Root
       ref={ref}
       className={cn(
-        'relative h-2 w-full overflow-hidden rounded-full bg-primary/20',
+        'bg-primary/20 relative h-2 w-full overflow-hidden rounded-full',
         className,
       )}
       {...props}
