+++ packages/app/src/lib/resources/blockstates.ts
@@ -1,13 +1,13 @@
 import { Mutex } from 'async-mutex'
 
 import fetcher from '@/fetcher'
+import { stripMinecraftPrefix } from '@/lib/utils'
 import { AssetFileInfosCache, useCacheStore } from '@/stores/cacheStore'
 import type {
   BlockStateApplyModelInfo,
   BlockStatesFile,
   BlockstatesData,
 } from '@/types'
-import { stripMinecraftPrefix } from '@/utils'
 
 import { getLogger } from '../logger'
