+++ packages/app/src/lib/resources/font/bitmap.ts
@@ -2,9 +2,9 @@ import { Mutex } from 'async-mutex'
 import { CanvasTexture, ImageLoader } from 'three'
 
 import fetcher from '@/fetcher'
+import { stripMinecraftPrefix } from '@/lib/utils'
 import { AssetFileInfosCache, useCacheStore } from '@/stores/cacheStore'
 import type { CDNFontProviderResponse } from '@/types'
-import { stripMinecraftPrefix } from '@/utils'
 
 const imageLoader = new ImageLoader()
 imageLoader.setCrossOrigin('anonymous')
