+++ packages/app/src/lib/resources/model.ts
@@ -1,9 +1,9 @@
 import { Mutex } from 'async-mutex'
 
 import fetcher from '@/fetcher'
+import { generateBuiltinItemModel, stripMinecraftPrefix } from '@/lib/utils'
 import { AssetFileInfosCache } from '@/stores/cacheStore'
 import type { ModelData, ModelElement, ModelFile } from '@/types'
-import { generateBuiltinItemModel, stripMinecraftPrefix } from '@/utils'
 
 import { getLogger } from '../logger'
